+++ database/migrations/2020_02_09_152805_create_template_styles_table.php
@@ -15,10 +15,10 @@ class CreateTemplateStylesTable extends Migration
     {
         Schema::create('template_styles', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('template_id')->unsigned()->nullable();
+            $table->unsignedBigInteger('template_id')->unsigned()->nullable();
             $table->foreign('template_id')->references('id')->on('templates')->onDelete('cascade');
 
-            $table->integer('style_id')->unsigned()->nullable();
+            $table->unsignedBigInteger('style_id')->unsigned()->nullable();
             $table->foreign('style_id')->references('id')->on('styles')->onDelete('cascade');
             $table->text('data')->nullable();
         });
